Update README.md
Pokemon Guessing Game
This is my first project using python! I have created a guessing game that is played in the terminal of your python IDE. It starts by asking for a name then asks if you would like to play the game. This question is inside a loop that allows for possible mistakes. If yes is entered the game continues, if no is entered the game ends. If neither is entered you will have 3 attempts to answer yes or no otherwise the game ends. 
The game asks 5 questions about the first generation of pokemon. The questions are accompanied by A-D answers. If the correct answer is inputted it adds one to the score variable and moves on to the next question. If the incorrect answer is inputted it moves onto the next question. After all 5 questions have been answered your score is printed with different commentary depending on how many questions you got right. And finally it prints the percentage you got right.

Thanks for reading and reviewing my code, hope you enjoy playing the game!